requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
opencv-python==4.10.0.84
jinja2==3.1.2
pillow==11.0.0
ultralytics==8.3.25
python-multipart==0.0.17